Conflicts:
	scilab/Scilab.sln
	scilab/Scilab_f2c.sln

+++ scilab/modules/xcos/src/java/org/scilab/modules/xcos/actions/DumpAction.java
@@ -23,6 +23,7 @@ import java.io.IOException;
 import org.apache.commons.logging.LogFactory;
 import org.scilab.modules.action_binding.highlevel.ScilabInterpreterManagement;
 import org.scilab.modules.action_binding.highlevel.ScilabInterpreterManagement.InterpreterException;
+import org.scilab.modules.graph.ScilabComponent;
 import org.scilab.modules.graph.ScilabGraph;
 import org.scilab.modules.graph.actions.base.DefaultAction;
 import org.scilab.modules.gui.menuitem.MenuItem;
@@ -82,10 +83,18 @@ public class DumpAction extends DefaultAction {
      */
     @Override
     public void actionPerformed(ActionEvent e) {
+        final XcosDiagram graph = (XcosDiagram) getGraph(e);
+
+        // action disabled when the cell is edited
+        final ScilabComponent comp = ((ScilabComponent) graph.getAsComponent());
+        if (comp.isEditing()) {
+            return;
+        }
+        
         try {
             String temp = FileUtils.createTempFile();
             new File(temp).deleteOnExit();
-            ((XcosDiagram) getGraph(e)).dumpToHdf5File(temp);
+            graph.dumpToHdf5File(temp);
             try {
                 String cmd = buildCall("import_from_hdf5", temp);
                 cmd += buildCall("deletefile", temp);
